test: Refactored test runner and BaseAPITestCase [#592]
+ Restructured tests for account app

+ Added .coveragerc
