+++ django_api/.coveragerc
@@ -1,0 +1,29 @@
+# .coveragerc to control coverage.py
+[run]
+omit =
+    # Django migration files
+    */migrations/*
+    */tests/*
+    */urls.py
+    */__init__.py
+    */settings/*
+
+[report]
+# Regexes for lines to exclude from consideration
+exclude_lines =
+    # Have to re-enable the standard pragma
+    pragma: no cover
+
+    # Don't complain about missing debug-only code:
+    def __repr__
+    if self\.debug
+
+    # Don't complain if tests don't hit defensive assertion code:
+    raise AssertionError
+    raise NotImplementedError
+
+    # Don't complain if non-runnable code isn't run:
+    if 0:
+    if __name__ == .__main__.:
+
+ignore_errors = True
